requirements_grpo.txt
langchain 
langchain-openai
google-generativeai
langchain-google-genai
python-dotenv